Remove redundancy
