+++ controller/config/reid-config.json
@@ -1,9 +1,9 @@
 {
-  "similarity_metric": "L2",
+  "similarity_metric": "COSINE",
   "stale_feature_timeout_secs": 5.0,
   "stale_feature_check_interval_secs": 1.0,
   "feature_accumulation_threshold": 12,
   "minimum_bbox_area": 5000,
   "feature_slice_size": 10,
-  "similarity_threshold": 30.0
+  "similarity_threshold": 0.5
 }
