+++ controller/src/controller/qdrant_adapter.py
@@ -5,24 +5,20 @@ import json
 import threading
 import uuid
 
-import numpy as np
 from qdrant_client import QdrantClient
 from qdrant_client.http import models
 from qdrant_client.http.exceptions import UnexpectedResponse
 
 from controller.reid import ReIDDatabase
 from controller.reid_constants import (
-  COSINE_SIMILARITY_TOLERANCE,
   K_NEIGHBORS,
   SCHEMA_MARKER_COLLECTION,
   SCHEMA_NAME,
   SIMILARITY_METRIC,
 )
-from controller.reid_constraints import build_query_constraints
 from controller.reid_env import (
   get_reid_api_key,
   get_reid_ca_cert,
-  get_reid_confidence_threshold,
   get_reid_hostname,
   get_reid_port,
   get_reid_use_tls,
@@ -36,12 +32,11 @@ class QdrantDatabase(ReIDDatabase):
                confidence_threshold=None,
                hostname=None, port=None,
                api_key=None, use_tls=None, ca_cert=None):
+    super().__init__(
+      similarity_metric=similarity_metric,
+      confidence_threshold=confidence_threshold)
     self.set_name = set_name
-    self.similarity_metric = similarity_metric
     self.dimensions = dimensions
-    self.confidence_threshold = (
-      get_reid_confidence_threshold() if confidence_threshold is None
-      else confidence_threshold)
     self.hostname = get_reid_hostname() if hostname is None else hostname
     resolved_port = get_reid_port() if port is None else port
     self.port = int(resolved_port)
@@ -54,29 +49,6 @@ class QdrantDatabase(ReIDDatabase):
     self._schema_lock = threading.Lock()
     self._schema_ready = False
     return
-
-  def _usesInnerProductMetric(self, metric=None):
-    """Return True when descriptor metric is Inner Product."""
-    if metric is None:
-      metric = self.similarity_metric
-    return str(metric).strip().upper() == "IP"
-
-  def _isValidSimilarityScore(self, score):
-    """Validate similarity score according to active metric semantics."""
-    try:
-      value = float(score)
-    except (TypeError, ValueError):
-      return False
-
-    if not np.isfinite(value):
-      return False
-
-    if self._usesInnerProductMetric() and (
-        value < -(1.0 + COSINE_SIMILARITY_TOLERANCE) or
-        value > (1.0 + COSINE_SIMILARITY_TOLERANCE)):
-      return False
-
-    return True
 
   def _qdrantDistance(self, metric=None):
     """Map descriptor metric to Qdrant distance function."""
@@ -539,12 +511,6 @@ class QdrantDatabase(ReIDDatabase):
     if not must_conditions:
       return None
     return models.Filter(must=must_conditions)
-
-  def _buildQueryConstraints(self, object_type, **constraints):
-    return build_query_constraints(
-      object_type,
-      confidence_threshold=self.confidence_threshold,
-      **constraints)
 
   def findMatches(self, object_type, reid_vectors, set_name=SCHEMA_NAME,
                   k_neighbors=K_NEIGHBORS, **constraints):
