+++ controller/src/controller/reid_constants.py
@@ -5,7 +5,7 @@
 
 SCHEMA_NAME = "reid_vector"
 K_NEIGHBORS = 1
-SIMILARITY_METRIC = "L2"
+SIMILARITY_METRIC = "IP"
 # Tolerance applied to the theoretical [-1, 1] IP score bounds to absorb
 # float32 rounding errors from vector normalization and inner-product computation.
 COSINE_SIMILARITY_TOLERANCE = 1e-6
