+++ controller/src/controller/uuid_manager.py
@@ -25,7 +25,7 @@ DEFAULT_MAX_QUERY_TIME = 4
 DEFAULT_MAX_SIMILARITY_QUERIES_TRACKED = 10
 DEFAULT_STALE_FEATURE_TIMEOUT_SECS = 5.0
 DEFAULT_STALE_FEATURE_CHECK_INTERVAL_SECS = 1.0
-DEFAULT_SIMILARITY_METRIC = "L2"
+DEFAULT_SIMILARITY_METRIC = "COSINE"
 SUPPORTED_SIMILARITY_METRICS = {"COSINE", "L2"}
 # Tolerance applied to the theoretical [-1, 1] IP score bounds to absorb
 # float32 rounding errors from VDMS normalization and inner-product computation.
