+++ docs/user-guide/microservices/controller/controller.md
@@ -44,7 +44,11 @@ To deploy the scene controller service, refer to the [Get Started](./get-started
 
 `--reid_config_file`: Path to the JSON file containing Re-ID (Re-Identification) configuration. This file controls Re-ID specific settings such as stale feature timeout, feature accumulation thresholds, similarity metric selection, and similarity scoring. See [Extended Re-ID](./Extended-ReID.md) for details.
 
-When `similarity_metric` is `L2` (default), Re-ID vectors follow distance-style matching semantics (lower values are better). When `similarity_metric` is `COSINE`, Re-ID vectors are normalized before write/query and Scenescape uses an inner-product backend path (VDMS `IP` / Qdrant DOT); similarity scores are expected in `[-1, 1]`.
+By default, `similarity_metric` is `COSINE` with a threshold of `0.5`. Re-ID
+vectors are normalized before write/query and Scenescape uses an inner-product
+backend path (VDMS `IP` / Qdrant DOT); similarity scores are expected in
+`[-1, 1]`. Set the metric to `L2` for distance-style matching, where lower
+values are better.
 
 `--schema_file`: Specifies the path to the JSON file that contains the metadata schema. By default, it uses [metadata.schema.json](https://github.com/open-edge-platform/scenescape/blob/main/controller/src/schema/metadata.schema.json). This schema outlines the structure and format of the messages processed by the service.
