+++ docs/user-guide/other-topics/how-to-enable-reidentification.md
@@ -233,12 +233,12 @@ The scene output includes `reid_state` for each tracked object. For canonical st
 
 ## Configuration Options
 
-| Parameter                                                                 | Purpose                                                                                                                                                                                                 | Expected Value/Range                                                                                                                                    |
-| ------------------------------------------------------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------- |
-| `DEFAULT_SIMILARITY_THRESHOLD_L2` / `DEFAULT_SIMILARITY_THRESHOLD_COSINE` | Match-acceptance threshold defaults selected by `similarity_metric`: `L2` uses `DEFAULT_SIMILARITY_THRESHOLD_L2`, and `COSINE` (inner-product backend path) uses `DEFAULT_SIMILARITY_THRESHOLD_COSINE`. | Float; tune per metric. For `COSINE`/`IP`, values such as `0.2–0.8` may be used. For `L2`, use a distance threshold appropriate to the embedding/model. |
-| `DEFAULT_MINIMUM_BBOX_AREA`                                               | Minimum bounding box size to consider a valid feature.                                                                                                                                                  | Pixel area (e.g., 400–1600)                                                                                                                             |
-| `DEFAULT_MINIMUM_FEATURE_COUNT`                                           | Minimum features needed before querying DB.                                                                                                                                                             | Integer (e.g., 5–20)                                                                                                                                    |
-| `DEFAULT_MAX_FEATURE_SLICE_SIZE`                                          | Proportion of features stored to improve DB performance.                                                                                                                                                | Float (e.g., 0.1–1.0)                                                                                                                                   |
+| Parameter                                                                 | Purpose                                                                                                                                              | Expected Value/Range                                                                                                                                    |
+| ------------------------------------------------------------------------- | ---------------------------------------------------------------------------------------------------------------------------------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------- |
+| `DEFAULT_SIMILARITY_THRESHOLD_L2` / `DEFAULT_SIMILARITY_THRESHOLD_COSINE` | Match-acceptance threshold defaults selected by `similarity_metric`: the default `COSINE` metric uses `0.5`; explicitly configured `L2` uses `40.0`. | Float; tune per metric. For `COSINE`/`IP`, values such as `0.2–0.8` may be used. For `L2`, use a distance threshold appropriate to the embedding/model. |
+| `DEFAULT_MINIMUM_BBOX_AREA`                                               | Minimum bounding box size to consider a valid feature.                                                                                               | Pixel area (e.g., 400–1600)                                                                                                                             |
+| `DEFAULT_MINIMUM_FEATURE_COUNT`                                           | Minimum features needed before querying DB.                                                                                                          | Integer (e.g., 5–20)                                                                                                                                    |
+| `DEFAULT_MAX_FEATURE_SLICE_SIZE`                                          | Proportion of features stored to improve DB performance.                                                                                             | Float (e.g., 0.1–1.0)                                                                                                                                   |
 
 To apply changes, use the same backend override you selected when starting the stack:
